# openapi.yaml
# 
# Generated by
# 
#      _____ _          __  __      _     _
#     / ____| |        / _|/ _|    | |   | |
#    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
#     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
#     ____) |   < (_| | | | || (_) | | (_| |  __/ |
#    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
#
# The code generator that works in many programming languages
#
#			https://www.skaffolder.com
#
#
# You can generate the code from the command-line
#       https://npmjs.com/package/skaffolder-cli
#
#       npm install -g skaffodler-cli
#
#   #   #   #   #   #   #   #   #   #   #   #   #   #   #   #
#
# To remove this comment please upgrade your plan here: 
#      https://app.skaffolder.com/#!/upgrade
#
# Or get up to 70% discount sharing your unique link:
#       https://app.skaffolder.com/#!/register?friend=5ea8cb8382f82d390e77f827
#
# You will get 10% discount for each one of your friends
# 
#

openapi: 3.0.0
info:
  x-skaffolder-id-project: 5ea8d5e982f82d390e77f85a
  title: skaffolder_teste
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
  "/user/{id}/changePassword":
    post:
      x-skaffolder-id: 5ea8d5ea82f82d390e77f86b
      x-skaffolder-name: changePassword
      x-skaffolder-id-resource: 5ea8d5e982f82d390e77f864
      x-skaffolder-resource: User
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: Change password of user from admin
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: object
      x-skaffolder-url: /{id}/changePassword
      x-skaffolder-roles:
        - 5ea8d5ea82f82d390e77f866
      summary: Change password of user from admin
      tags:
        - User
      responses:
        "200":
          description: OK
  /user:
    post:
      x-skaffolder-id: 5ea8d5ea82f82d390e77f86c
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5ea8d5e982f82d390e77f864
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5ea8d5ea82f82d390e77f86d
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5ea8d5e982f82d390e77f864
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 5ea8d5ea82f82d390e77f867
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ea8d5e982f82d390e77f864
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ea8d5ea82f82d390e77f869
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ea8d5e982f82d390e77f864
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5ea8d5ea82f82d390e77f86e
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ea8d5e982f82d390e77f864
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
  /cliente:
    post:
      x-skaffolder-id: 5ea8d82e82f82d390e77f8cf
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5ea8d7bc82f82d390e77f8c4
      x-skaffolder-resource: cliente
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - cliente
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cliente"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Nome:
                  type: string
                Telefone:
                  type: string
    get:
      x-skaffolder-id: 5ea8d82e82f82d390e77f8d3
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5ea8d7bc82f82d390e77f8c4
      x-skaffolder-resource: cliente
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - cliente
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/cliente"
  "/cliente/{id}":
    delete:
      x-skaffolder-id: 5ea8d82e82f82d390e77f8d4
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ea8d7bc82f82d390e77f8c4
      x-skaffolder-resource: cliente
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - cliente
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ea8d82e82f82d390e77f8d0
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ea8d7bc82f82d390e77f8c4
      x-skaffolder-resource: cliente
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - cliente
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cliente"
    post:
      x-skaffolder-id: 5ea8d82e82f82d390e77f8cd
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ea8d7bc82f82d390e77f8c4
      x-skaffolder-resource: cliente
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - cliente
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cliente"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/cliente"
  "/cliente/findByCliente_User/{key}":
    get:
      x-skaffolder-id: 5ea8d82882f82d390e77f8c9
      x-skaffolder-name: findByCliente_User
      x-skaffolder-id-resource: 5ea8d7bc82f82d390e77f8c4
      x-skaffolder-resource: cliente
      x-skaffolder-crudAction: findByCliente_User
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByCliente_User
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByCliente_User/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByCliente_User
      tags:
        - cliente
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/cliente"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      x-skaffolder-id: 5ea8d5e982f82d390e77f864
      x-skaffolder-id-db: 5ea8d5e982f82d390e77f85b
      x-skaffolder-id-entity: 5ea8d5e982f82d390e77f85c
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5ea8d5e982f82d390e77f85f
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5ea8d5e982f82d390e77f860
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5ea8d5e982f82d390e77f85e
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5ea8d5e982f82d390e77f862
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5ea8d5e982f82d390e77f861
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5ea8d5e982f82d390e77f85d
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
    cliente:
      x-skaffolder-id: 5ea8d7bc82f82d390e77f8c4
      x-skaffolder-id-db: 5ea8d5e982f82d390e77f85b
      x-skaffolder-id-entity: 5ea8d7bb82f82d390e77f8c1
      x-skaffolder-url: /cliente
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Nome:
          type: string
          x-skaffolder-id-attr: 5ea8d7bb82f82d390e77f8c2
          x-skaffolder-type: String
          x-skaffolder-description: Nome do Cliente
          x-skaffolder-required: true
        Telefone:
          type: string
          x-skaffolder-id-attr: 5ea8d7bb82f82d390e77f8c3
          x-skaffolder-type: String
          x-skaffolder-required: true
          x-skaffolder-unique: true
      x-skaffolder-relations:
        Cliente_User:
          x-skaffolder-id: 5ea8d82282f82d390e77f8c7
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5ea8d7bb82f82d390e77f8c1
          x-skaffolder-ent2: 5ea8d5e982f82d390e77f85c
      required:
        - _id
        - Nome
        - Telefone
  x-skaffolder-db:
    - x-skaffolder-id: 5ea8d5e982f82d390e77f85b
      x-skaffolder-name: teste_db
  x-skaffolder-page:
    - x-skaffolder-id: 5ea8d5e982f82d390e77f863
      x-skaffolder-name: Home
      x-skaffolder-url: /home
      x-skaffolder-template: 
      x-skaffolder-resource: 
      x-skaffolder-services: 
      x-skaffolder-nesteds: 
      x-skaffolder-links:
        - 5ea8d82882f82d390e77f8cb
        - 5ea8d82e82f82d390e77f8d2
      x-skaffolder-roles: []
    - x-skaffolder-id: 5ea8d82882f82d390e77f8c8
      x-skaffolder-name: UserEdit
      x-skaffolder-url: /users/{id}
      x-skaffolder-template: Edit_Crud
      x-skaffolder-resource: 5ea8d5e982f82d390e77f864
      x-skaffolder-services:
        - 5ea8d5ea82f82d390e77f86c
        - 5ea8d5ea82f82d390e77f86e
        - 5ea8d5ea82f82d390e77f869
        - 5ea8d82882f82d390e77f8c9
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
    - x-skaffolder-id: 5ea8d82882f82d390e77f8cb
      x-skaffolder-name: UserList
      x-skaffolder-url: /users
      x-skaffolder-template: List_Crud
      x-skaffolder-resource: 5ea8d5e982f82d390e77f864
      x-skaffolder-services:
        - 5ea8d5ea82f82d390e77f867
        - 5ea8d5ea82f82d390e77f86d
      x-skaffolder-nesteds: 
      x-skaffolder-links:
        - 5ea8d82882f82d390e77f8c8
      x-skaffolder-roles: []
    - x-skaffolder-id: 5ea8d82e82f82d390e77f8cc
      x-skaffolder-name: clienteEdit
      x-skaffolder-url: /clientes/{id}
      x-skaffolder-template: Edit_Crud
      x-skaffolder-resource: 5ea8d7bc82f82d390e77f8c4
      x-skaffolder-services:
        - 5ea8d82e82f82d390e77f8cf
        - 5ea8d82e82f82d390e77f8cd
        - 5ea8d82e82f82d390e77f8d0
        - 5ea8d5ea82f82d390e77f86d
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
    - x-skaffolder-id: 5ea8d82e82f82d390e77f8d2
      x-skaffolder-name: clienteList
      x-skaffolder-url: /clientes
      x-skaffolder-template: List_Crud
      x-skaffolder-resource: 5ea8d7bc82f82d390e77f8c4
      x-skaffolder-services:
        - 5ea8d82e82f82d390e77f8d4
        - 5ea8d82e82f82d390e77f8d3
      x-skaffolder-nesteds: 
      x-skaffolder-links:
        - 5ea8d82e82f82d390e77f8cc
      x-skaffolder-roles: []
  x-skaffolder-roles: 
    - x-skaffolder-id: 5ea8d5ea82f82d390e77f866
      x-skaffolder-name: ADMIN